Fix modifier with return values needing phis
When the return value is stored in a local variable, then phis might be
needed.
